+++ src/components/TxHistory/TxHistory.js
@@ -64,6 +64,7 @@ const AvailableAmountBanner = withTranslations(getTranslations)(
       text={`${formatAda(amount)} ${
         translations.availableAmountBanner.ADASymbol
       }`}
+      boldText
     />
   ),
 )
