+++ src/components/UiKit/styles/Banner.style.js
@@ -1,0 +1,23 @@
+// @flow
+
+import {StyleSheet} from 'react-native'
+
+const styles = StyleSheet.create({
+  banner: {
+    backgroundColor: '#F0F3F5',
+    padding: 16,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  textError: {
+    color: '#FF1351',
+  },
+  label: {
+    marginBottom: 6,
+  },
+  bold: {
+    fontWeight: 'bold',
+  },
+})
+
+export default styles
